Add files via upload
"(COMMIT OF ORIGINAL .m FILE: original commit before edit failed, but here is the description): This simulation tracks the depolymerization of a filament by calculating the time for a proetin to land on a random position, the time that protein takes to reach either end (based on the Gambler's Ruin model), then the time it takes for the protein to depolymerize the filament. It does this repeatedly over the course of many runs, then averages the time for the filament to depolymerize. Right now, the measurements are done in monomers, but in a future edit, I will adjust it to calculate based of of filament length in microns and protein concentration in nM."

SECOND (CURRENT) COMMIT OF FILES: "These are the runs of the Gambler's Ruin model using the parameters from the Helenius paper, scaled to a monomer size of .128 microns. These are semi-random parameters, and only used because they were readily available and I wanted to test the simulation. The data from this simulation should be compared to the Gillespie single protein simulation, not the multiple protein simulation."