Add files via upload
Changed the simulation so that, instead of the walk step taking up the average amount of time a protein would take to reach EITHER end, the simulation chooses whether it is going left or right (based on probability), then uses that amount of time as the walk step time. Before, the simulation results did not match the Gillespie single protein simulation results, but now they do.